% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHPT_ASPT_Score.R
\name{WHPT_ASPT_Score}
\alias{WHPT_ASPT_Score}
\title{WHPT-ASPT Score Calculator}
\usage{
WHPT_ASPT_Score(x, Sample_Label = NULL)
}
\arguments{
\item{x}{Benthic Macroinvertebrate dataframe (csv or data.frame).}

\item{Sample_ID}{Name of column containing Sample Name.}
}
\description{
Returns the WHPT-ASPT score for each sample submitted.
}
